Changes to fighters, items and the SOUL

+++ fighters/characters/character.gd
@@ -1,33 +1,24 @@
 extends AbstractFighter
 class_name Character
 
-const CHECK_ACT = preload("res://fighters/characters/acts/check/check.tres")
+const CHECK_ACT := preload("res://fighters/characters/acts/check/check.tres")
 
 enum Animations {
 	IDLE, PREP_ATTACK, PREP_ACT, PREP_ITEM, PREP_SPARE, ATTACK, ACT, USE_ITEM, SPARE, DEFEND, FAINT, REVIVE
 }
 
-@export var magic := 0:
-	get():
-		var equipped_magic := 0 if !weapon else weapon.magic
-		for armor: Equippable in armors:
-			equipped_magic += armor.magic
-		return magic + equipped_magic
 @export var uses_magic := false
 @export_color_no_alpha var main_color := Color.WHITE
 @export_color_no_alpha var icon_color := Color.GRAY
 @export var icon: Texture2D = preload("res://ui/battle/char_menu/res/sample_char_icon.png")
+@export var willNotEquip : Array[Equippable.Category]
 
-@export_node_path("Sprite2D") var main_sprite
-var sprite: Sprite2D
-var mat: ShaderMaterial
 var alive := true
 var defending := false
 var shake := 0.0
 
 ## The list of default spells the character can use. Only for characters with do_magic set to true.
 @export var spells: Array[Spell] = []
-
 
 signal act_finished
 signal spell_finished
@@ -36,13 +27,10 @@ signal spare_finished
 signal faint_finished
 
 func _ready() -> void:
-	if !main_sprite:
-		return
-	sprite = get_node(main_sprite)
-	mat = ShaderMaterial.new()
-	mat.shader = preload("res://fighters/characters/generic_character.gdshader")
-	sprite.material = mat
-	await get_tree().create_timer(randf_range(0.0, 0.6)).timeout
+	create_shader("generic_character")
+
+func create_shader(shaderName : String) -> void:
+	super.create_shader(shaderName)
 	do_animation(Animations.IDLE)
 
 func swap_armor(p_id: int, p_armor: Equippable) -> void:
@@ -102,17 +90,14 @@ func use_item(p_character: Character, p_item: int) -> void:
 		await Global.text_finished
 		item_used.emit()
 		return
-	match item.type:
-		Item.TYPE.NONE:
-			Global.display_text.emit("  * " + title + " used the " + item.name + ".", true)
-			await Global.text_finished
-			await get_tree().create_timer(0.01).timeout
-			Global.display_text.emit("  * But nothing happened...", true)
-		Item.TYPE.HEAL:
-			p_character.heal(item.amount)
-			Global.display_text.emit("  * " + title + " used the " + item.name + "!", true)
-			Global.delete_item(p_item)
-	await Global.text_finished
+	var predicate : ItemUsePredicate
+	if item.usePredicate == null:
+		predicate = ItemUsePredicate.new()
+	else:
+		predicate = item.usePredicate
+	await predicate.use_item(self, p_character)
+	if item.usePredicate.remove_after_use:
+		Global.delete_item(p_item)
 	do_animation(Animations.IDLE)
 	item_used.emit()
 
@@ -145,14 +130,14 @@ func faint() -> void:
 	faint_finished.emit()
 
 func revive() -> void:
-	if current_hp < max_hp * 0.17:
-		current_hp = ceili(max_hp * 0.17)
+	if current_hp < get_value_from_stat(Stats.MAX_HP) * 0.17:
+		current_hp = ceili(get_max_hp() * 0.17)
 	alive = true
 	do_animation(Animations.REVIVE)
 
 func hurt(p_damage: int) -> void:
 	shake_sprite(4.0)
-	p_damage = int(maxi(1, p_damage - 3 * defense) * (1.0 if !defending else 2.0 / 3.0))
+	p_damage = int(maxi(1, p_damage - 3 * get_defense()) * (1.0 if !defending else 2.0 / 3.0))
 	current_hp -= p_damage
 	health_changed.emit(current_hp)
 	create_text(str(p_damage), Color.WHITE)
@@ -165,8 +150,8 @@ func heal(p_amount: int) -> void:
 	Sounds.play("snd_power")
 	if !alive and current_hp >= 0:
 		revive()
-	if current_hp >= max_hp:
-		current_hp = max_hp
+	if current_hp >= get_max_hp():
+		current_hp = get_max_hp()
 		create_text("MAX", Global.GREEN)
 	else:
 		create_text(str(p_amount), icon_color)
